+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add `Ezcater/RspecRequireFeatureFlagMock` cop.
 - Add `Ezcater/RspecDotNotSelfDot` cop.
 - Add `Ezcater/StyleDig` cop.
+- Add `Ezcater/RspecRequireBrowserMock` cop.